+++ math/src/elliptic_curve/curves/mod.rs
@@ -1,0 +1,1 @@
+pub mod example;
